+++ InterviewQuestions/Square.cpp
@@ -19,10 +19,10 @@ Square::~Square()
 
 bool Square::IsOverlappig(Square &other)
 {
-	bool result1 = other.x >= x && other.y >= y && other.x <= (x + width) && other.y <= (y + height); // other's top left falls within this area
-	bool result2 = other.x >= x && other.y <= y && other.x <= (x + width) && (other.y + other.height) <= (y + height); // other's bottom left falls within this area
-	bool result3 = other.x <= x && other.y >= y && (other.x + other.width) <= (x + width) && other.y <= (y + height); // other's top right falls within this area
-	bool result4 = other.x <= x && other.y <= y && (other.x + other.width) >= x && (other.y + other.height) >= y; // other's bottom right falls within this area
+	bool result1 = other.x >= x && other.y >= y && other.x <= (long)(x + width) && other.y <= (long)(y + height); // other's top left falls within this area
+	bool result2 = other.x >= x && other.y <= y && other.x <= (long)(x + width) && (long)(other.y + other.height) <= (long)(y + height); // other's bottom left falls within this area
+	bool result3 = other.x <= x && other.y >= y && (long)(other.x + other.width) <= (long)(x + width) && other.y <= (long)(y + height); // other's top right falls within this area
+	bool result4 = other.x <= x && other.y <= y && (long)(other.x + other.width) >= x && (long)(other.y + other.height) >= y; // other's bottom right falls within this area
 	return result1 | result2 | result3 | result4;
 }
